Added API for booking
Added API for booking and manage functions between files
